v2/metadata_service/requirements.txt
fastapi
uvicorn
python-multipart
motor
passlib[argon2]
python-jose[cryptography]
python-dotenv
python-dateutil
httpx
shamirs
pynacl